Add discovery_metadata table to PSQL and index on ES.
Add discovery_metadata table to PSQL and index on Elasticsearch. The metadata initializes for the 11 datasets on the `admin init` command, and updates the levels and spatiotemporal extents on the `ingest` and `uv-index generate` commands. The metadata is a geoJSON string in the PSQL table and inserted into the Elasticsearch index with individual fields.